+++ solvis/dochelper/extension.py
@@ -1,7 +1,8 @@
-"""
-A griffe extension lifted from https://mkdocstrings.github.io/griffe/guide/users/extending/#full-example
+"""A griffe extension.
 
+Lifted from https://mkdocstrings.github.io/griffe/guide/users/extending/#full-example
 """
+
 import ast
 import inspect
 from typing import Optional, Union
@@ -14,6 +15,8 @@ logger.setLevel('DEBUG')
 
 
 class DynamicDocstrings(griffe.Extension):
+    """The helper class."""
+
     def __init__(self, object_paths: Optional[list[str]] = None) -> None:
         self.object_paths = object_paths
         logger.debug(object_paths)
@@ -52,10 +55,14 @@ class DynamicDocstrings(griffe.Extension):
             logger.debug(f"Object {obj.path} does not have a __doc__ attribute")
             return
 
+        # do not primitive class attributes
+        if isinstance(runtime_obj, (str, int, list)):
+            return
+
         # Update the object instance with the evaluated docstring.
         try:
             docstring = inspect.cleandoc(docstring)
-            logger.warning(docstring)
+            logger.warning(f"evaluated docstring: {docstring}")
         except AttributeError:
             return
 
@@ -70,3 +77,19 @@ class DynamicDocstrings(griffe.Extension):
                 parser_options=agent.docstring_options,
             )
         logger.info(f'updated docstring for {obj.path}: {docstring}')
+        # if obj.path == 'solvis.solution.inversion_solution.inversion_solution_file.InversionSolutionFile.RATES_PATH':
+        #     ds = griffe.Docstring(
+        #         docstring,
+        #         parent=obj,
+        #         parser=agent.docstring_parser,
+        #         parser_options=agent.docstring_options,
+        #     )
+
+        #     print(dir(ds))
+        #     print(dir(ds.parsed[0]))
+        #     print(ds.parsed[0].as_dict())
+        #     print(ds.lines)
+
+        #     print()
+
+        #     assert 0