+++ solvis/solution/dataframe_models.py
@@ -1,11 +1,10 @@
 """This module defines the dataframes returned by solvis.
 
- - We use the `panderas` library to define the dataframe schemas and solvis testing validates that
+- We use the `panderas` library to define the dataframe schemas and solvis testing validates that
    dataframes have the expected columns names and types.
- - currenty schema attributes are NOT the actual dataframe column names, For these see the
+- currenty schema attributes are NOT the actual dataframe column names, For these see the
     code and the `Field(alias={column_name}) value. See below...
- - refactoring of column names is expected in a future release, but legacy column names will be supported.
-
+- refactoring of column names is expected in a future release, but legacy column names will be supported.
 """
 
 from typing import TYPE_CHECKING
@@ -52,23 +51,6 @@ class FaultSectionSchemaBase(pda.DataFrameModel):
     """A base schema for common dataframe columns.
 
     This is not used directly, but it's fields are reused in other schema.
-
-    Attributes:
-     index: unique index on section_id.
-     section_id: the section id (OpenSHA:FaultID).
-     section_name: section name (OpenSHA:FaultName).
-     dip_angle: dip angle in degrees (OpenSHA:DipDeg).
-     rake: rake angle in degrees (OpenSHA:Rake).
-     lower_depth: fault lower depth in meters (OpenSHA:LowDepth).
-     upper_depth: fault upper depth in metres (OpenSHA:UpDepth).
-     aseismic_slip_factor: ratio of aseismic slip  (OpenSHA:AseismicSlipFactor).
-     coupling_coefficient: coupling coefficient (OpenSHA:CouplingCoeff).
-     parent_fault_id: the parent fault id (OpenSHA:ParentID).
-     parent_fault_name: the parent fault name (OpenSHA:ParentName).
-     dir_direction: dip direction in degrees  = (OpenSHA:DipDir).
-     geometry: a `shapely.geometry` object representing the section surface.
-     target_slip_rate: a target slip rate (OpenSHA:'Target Slip Rate').
-     target_slip_rate_stddev: standard deviation for the target slip rate (OpenSHA:'Target Slip Rate StdDev').
     """
 
     section_id: Series[pd.Int32Dtype] = pda.Field(alias='FaultID')
@@ -88,10 +70,7 @@ class FaultSectionSchemaBase(pda.DataFrameModel):
 
 
 class FaultSectionSchema(FaultSectionSchemaBase):
-    """A Dataframe schema for `fault_section`
-
-    Attributes:
-     index: unique index on section_id.
+    """A Dataframe schema for `fault_section`.
 
     Note:
      remaining attributes are inherited from FaultSectionSchemaBase
@@ -104,14 +83,10 @@ class FaultSectionSchema(FaultSectionSchemaBase):
 
 
 class FaultSectionWithSolutionSlipRate(FaultSectionSchemaBase):
-    """A Dataframe schema for `fault_sections_with_solution_slip_rates`
+    """A Dataframe schema for `fault_sections_with_solution_slip_rates`.
 
     Notes:
      - this just adds one column to FaultSectionSchemaBase/FaultSectionSchema
-
-    Attributes:
-     index: unique index.
-     solution_slip_rate: # TODO: what is this ??
     """
 
     index: Index[pd.Int64Dtype] = pda.Field(alias='section_id')
@@ -119,10 +94,10 @@ class FaultSectionWithSolutionSlipRate(FaultSectionSchemaBase):
 
 
 class FaultSectionRuptureRateSchema(FaultSectionSchemaBase):
-    """A Dataframe schema for `fault_sections_with_rupture_rates`
+    """A Dataframe schema for `fault_sections_with_rupture_rates`.
 
     This table joins all permutations of rupture_id and section_id
-     with the aggregate rupture columns.
+    with the aggregate rupture columns.
 
     Todo:
      - why is this schema used in inversion_solution_file, shouldn't it exclusivly on fault_system_solution ??
@@ -173,7 +148,7 @@ class FaultSectionRuptureRateSchema(FaultSectionSchemaBase):
 
 
 class RuptureSectionSchema(pda.DataFrameModel):
-    """A Dataframe schema for `rupture_section`
+    """A Dataframe schema for `rupture_section`.
 
     This is a `join` table iterating all permutations of rupture_id and section_id.
 
@@ -192,6 +167,8 @@ class RuptureSectionSchema(pda.DataFrameModel):
 
 
 class RuptureBaseSchema(pda.DataFrameModel):
+    """A Dataframe schema base."""
+
     magnitude: Series[pd.Float32Dtype] = pda.Field(alias='Magnitude')
     mean_rake: Series[pd.Float32Dtype] = pda.Field(alias='Average Rake (degrees)')
     area: Series[pd.Float32Dtype] = pda.Field(alias='Area (m^2)')
@@ -199,11 +176,7 @@ class RuptureBaseSchema(pda.DataFrameModel):
 
 
 class RupturesWithRuptureRatesSchema(RuptureBaseSchema):
-    """A Dataframe schema for `InversionSolution.ruptures_with_rupture_rates`
-
-    Attributes:
-
-    """
+    """A Dataframe schema for `InversionSolution.ruptures_with_rupture_rates`."""
 
     class Config:
         strict = True
@@ -214,11 +187,7 @@ class RupturesWithRuptureRatesSchema(RuptureBaseSchema):
 
 
 class FSS_RupturesWithRuptureRatesSchema(RuptureBaseSchema):
-    """A Dataframe schema for `FaultSystemSolution.ruptures_with_rupture_rates`
-
-    Attributes:
-
-    """
+    """A Dataframe schema for `FaultSystemSolution.ruptures_with_rupture_rates`."""
 
     class Config:
         strict = True
@@ -238,7 +207,7 @@ class FSS_RupturesWithRuptureRatesSchema(RuptureBaseSchema):
 
 
 class RuptureSectionsWithRuptureRatesSchema(RuptureBaseSchema):
-    """A Dataframe schema for `InversionSolution.rs_with_rupture_rates`"""
+    """A Dataframe schema for `InversionSolution.rs_with_rupture_rates`."""
 
     class Config:
         strict = True
@@ -252,7 +221,7 @@ class RuptureSectionsWithRuptureRatesSchema(RuptureBaseSchema):
 
 
 class FSS_RuptureSectionsWithRuptureRatesSchema(RuptureBaseSchema):
-    """A Dataframe schema for `FaultSystemSolution.rs_with_rupture_rates`"""
+    """A Dataframe schema for `FaultSystemSolution.rs_with_rupture_rates`."""
 
     class Config:
         strict = True
@@ -277,7 +246,7 @@ class FSS_RuptureSectionsWithRuptureRatesSchema(RuptureBaseSchema):
 
 ### Below are from InversionSolutionFile
 class RuptureRateSchema(pda.DataFrameModel):
-    """A Dataframe schema for `InversionSolutionFile.rupture_rates`
+    """A Dataframe schema for `InversionSolutionFile.rupture_rates`.
 
     Attributes:
      index: unique index on rupture_id.
@@ -294,6 +263,8 @@ class RuptureRateSchema(pda.DataFrameModel):
 
 
 class RuptureSchema(RuptureBaseSchema):
+    """A dataframe schema for Ruptures."""
+
     class Config:
         strict = True